Add plotting module
