+++ benchmark.py
@@ -7,9 +7,10 @@ import matplotlib.pyplot as plt
 import numpy as np
 import typer
 
+import src
+import src.plotting
 from src.constants import ProblemType
 from src.cp import CP_MODELS
-from src.plotting import plot_schedule
 from src.ProblemData import ProblemData
 
 
@@ -23,12 +24,12 @@ def main(
 ):
     for where in datas:
         path = Path(where)
-        num_jobs, num_stages, time, lb, ub, gap = main_(
+        num_jobs, num_stages, time, lb, ub, gap = solve(
             path, time_limit, problem_type.value, num_procs
         )
 
 
-def main_(
+def solve(
     path,
     time_limit,
     problem_type,
@@ -79,43 +80,16 @@ def CPLEX_CP_solve(
     lb = result.get_objective_bounds()[0]
     ub = result.get_objective_values()[0]
 
-    if problem_type == "Parallelmachine":
-        schedule = result2schedule(data, result)
-        plot_schedule(data, schedule)
+    if problem_type == "Unrelatedparallelmachines":
+        src.plotting.upmresult2plot(data, result)
 
     if problem_type == "Hybridflowshop":
         hfsresult2plot(data, result)
 
     if problem_type == "Complexdistributedflowshop":
-        dpfsresult2plot(data, result)
+        src.plotting.dpfsresult2plot(data, result)
 
     return lb, ub
-
-
-class DFSItem(NamedTuple):
-    job: int
-    stage: int
-    machine: int
-    start: int
-    duration: int
-
-
-def result2schedule(data, result):
-    schedule = []
-
-    for machine in range(data.num_machines):
-        job_start = []
-
-        for job in range(data.num_jobs):
-            var = result.get_var_solution(f"O_{job}_{machine}")
-
-            if not var.is_absent():
-                job_start.append((job, var.get_start()))
-
-        sequence = [job for (job, *_) in sorted(job_start, key=lambda x: x[1])]
-        schedule.append(sequence)
-
-    return schedule
 
 
 def hfsresult2schedules(data, result):
@@ -197,62 +171,6 @@ class Item(NamedTuple):
     machine: int
     start: int
     duration: int
-
-
-def dpfsresult2plot(data, result):
-    """
-    Extracts the schedule for the DPFS and plots it.
-    """
-    schedule = []
-
-    for i, k in product(range(data.num_machines), range(data.num_factories)):
-        # Only get the first machine because we assume permutation flow shops.
-        seq_var = result.get_var_solution(f"S_{i}_{k}")
-
-        for interval_var in seq_var.get_interval_variables():
-            name = interval_var.get_name()
-            job = int(name.split("_")[1])
-            item = DFSItem(
-                job,
-                k,
-                i,
-                interval_var.get_start(),
-                interval_var.get_length(),
-            )
-            schedule.append(item)
-
-    dpfs_plot(data, schedule)
-
-
-def dpfs_plot(data, schedule, ax=None):
-    _, ax = plt.subplots(data.num_factories, 1, figsize=(10, 10))
-
-    def get_completion_time(item):
-        return item.start + item.duration
-
-    # Defining custom 'xlim' and 'ylim' values.
-    latest = max(schedule, key=get_completion_time)
-    custom_xlim = (0, get_completion_time(latest))
-
-    # Setting the values for all axes.
-    plt.setp(ax, xlim=custom_xlim)
-
-    # Color per job
-    hsv = plt.get_cmap("hsv")
-    colors = hsv(np.linspace(0, 1.0, data.num_jobs))
-
-    for job, stage, machine, start, duration in schedule:
-        ax[stage].barh(machine, width=duration, left=start, color=colors[job])
-        ax[stage].text(start + duration / 4, machine, job, va="center")
-
-    for factory in range(data.num_factories):
-        ax[factory].set_ylabel(f"Factory {factory+1}")
-
-        # Invert y-axis to have the first machine at the top.
-        ax[factory].set_ylim(ax[factory].get_ylim()[::-1])
-
-    plt.xlabel("Time")
-    plt.show()
 
 
 if __name__ == "__main__":
