+++ src/plotting/__init__.py
@@ -1,0 +1,2 @@
+from .plot_dfs import dpfsresult2plot
+from .plot_upm import upmresult2plot
